Beacon Android Application
This Android Application works with Beacons to allow users to find the
list of items in a particular category of the store when the user goes
near that particular Beacon. All the details of the items like price,
discount etc are display using the Amazon Web Service where the
database is created and we get the details of the items in the form of
JSON file.