Beacons Android Application
This Android application allow user to get the list of items in a
particular category in a store where beacons are equipped. When the
user go near a particular category items, the beacons lets the app show
all the items present in that category and price, discount and other
details of the items are shown. We used amazon web services for storing
and updating the item list in the database, and this data is retrieved
in the form of JSON query which is processed in the application